requirements.txt
tqdm
numpy
matplotlib
Pillow
tensorflow==2.5.0
